added default category profile
